requirements.txt
grpcio==1.31.0
grpcio-reflection==1.31.0
grpcio-tools==1.31.0
mypy-protobuf==1.20
regex
